+++ proto/cosmos/staking/v1beta1/staking.proto
@@ -231,10 +231,10 @@ message UnbondingDelegationEntry {
   string balance = 4  [(gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int", (gogoproto.nullable) = false];
 
   // Incrementing id that uniquely identifies this entry
-  uint64 id = 5;
+  uint64 unbonding_op_id = 5;
 
   // True if this entry's unbonding has been stopped by an external module
-  bool on_hold = 6; 
+  bool unbonding_on_hold = 6; 
 }
 
 // RedelegationEntry defines a redelegation object with relevant metadata.
@@ -258,10 +258,10 @@ message RedelegationEntry {
       [(gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Dec", (gogoproto.nullable) = false];
 
   // Incrementing id that uniquely identifies this entry
-  uint64 id = 5;
+  uint64 unbonding_op_id = 5;
 
   // True if this entry's unbonding has been stopped by an external module
-  bool on_hold = 6; 
+  bool unbonding_on_hold = 6; 
 }
 
 // Redelegation contains the list of a particular delegator's redelegating bonds
